Move video to repository folder
