+++ src/services/repositories/video/video.service.ts
@@ -2,9 +2,9 @@ import { Model } from 'mongoose';
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Video } from './schemas/video.schema';
-import { CreateVideoDto } from './dto/create-video.dto';
 import { TrendingChainDto, trendingChainProjection } from './dto/trending-chain.dto';
-import { normaliseTags } from '../../utils/normaliseTags';
+import { normaliseTags } from '../../../utils/normaliseTags';
+import { DbVideoToPublishDto, dbVideoToPublishProjection } from './dto/videos-to-publish.dto';
 
 type TrendingChainResponse = Array<({ 
   permlink: string;
@@ -26,19 +26,14 @@ type TrendingChainResponse = Array<({
 export class VideoService {
   constructor(@InjectModel(Video.name) private videoModel: Model<Video>) {}
 
-  async create(createCatDto: CreateVideoDto): Promise<Video> {
-    const createdCat = new this.videoModel(createCatDto);
-    return createdCat.save();
-  }
-
   async getTrendingForChain(): Promise<TrendingChainResponse> {
     const limit = 50;
 
-     const videos: TrendingChainDto[] = await this.videoModel.find({ status: 'published' }, trendingChainProjection)
-      .sort({ score: -1 })
-      .limit(limit)
-      .lean()
-      .exec();
+    const videos = await this.videoModel.find<TrendingChainDto>({ status: 'published' }, trendingChainProjection)
+    .sort({ score: -1 })
+    .limit(limit)
+    .lean()
+    .exec();
 
     return videos.map((e ,i) => ({
       trending_position: i + 1,
